tests/wpt/tests/html/canvas/offscreen/text/2d.text.align.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.text.align.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.align.valid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.textAlign = 'start';
  _assertSame(ctx.textAlign, 'start', "ctx.textAlign", "'start'");

  ctx.textAlign = 'end';
  _assertSame(ctx.textAlign, 'end', "ctx.textAlign", "'end'");

  ctx.textAlign = 'left';
  _assertSame(ctx.textAlign, 'left', "ctx.textAlign", "'left'");

  ctx.textAlign = 'right';
  _assertSame(ctx.textAlign, 'right', "ctx.textAlign", "'right'");

  ctx.textAlign = 'center';
  _assertSame(ctx.textAlign, 'center', "ctx.textAlign", "'center'");
  t.done();

});
</script>
